Add AvatarProvider
